Navigation Update & Panel (#14)
* added new images and finished prologue part 1

* Rollback Phase 1

+Added logic for performing "rollback". Rollback is avaialbe when a
dialogue choice is present. Pressing the rollback button will revert
the chat back to the previous choice if one is avaiable. If not, the
chapter will reset.

+Updated the scene to include the new rollback button on the
MessagingConversationPanel

* Rollback Impromvents and Crash Fix

*Fixed an error that was occuring when attempting to rollback the
social media posts if the list was empty

*Rolling back will now open the correct character's conversation panel
where the next choice is present. This logic will trigger when the
system caused the a previously opened conversation panel to be remove.

* finished ep13

* chapter changes

* full update

* Rollback Error & Crash Fix

*Fixed an issue that was causing the rollback system to delete extra
dialogue lines if the rollback was performed in a DialogueChoiceNode
that had the "Has Character Input" flag set to false.

*Fixed an issue that was causing the game to crash when performing a
rollback to the very start of a conversation.

* Timelapse Issue

*Rollback system will now correctly handle removing timelapse dialogue
messages if they are present.

* Navigation Panel

+Added navigation panel with a "Home" and "Back" button to allow the
player to navigation between opened panel in a similar way it works on
mobile devices. The system uses a command stack to queue windows that
are open and allow "undo" to be performed to close them

*Updated the scene to accomodate the new navigation panel.

* UI Update

*Fixed an issue where opening an image via the Gallery app to view in
full screen would not correclty close when using the navigation panel.

*Removed close buttons from the various app and UI windows. Navigation
is now done via the navigation panel at the bottom of the screen

* Logcat Package

+Added Android Logcat to help debug specific issues with android builds

---------

Co-authored-by: Bubbles-del-cpu <dajofisser@gmail.com>