Notification Update (#16)
* Dialogue Node Update (Yellow Block)

*Yellow blocks of the dialogue editor can now accept and handle an
empty string in the "Displayed Text" field without creating an empty
dialogue message in the conversation window

*Updated the notification system to alter the notifcation text if a
yellow block contains an empty "Displayed Text" field. The system will
now check for an image and video if no text is found and alter the
notificiation text to say "has sent a new picture/video"

* Action Choices (Blue block update)

*Added logic for handling "action" style choices in the
DialogueChoiceNodes (blue blocks). These messages are prefixed with *
and when selected will not sent a message to the conversation window
but will move the dialogue tree forwards

*Fixed an issue where the header for the messaging conversations was
not being reset when exiting the conversation.

*The overlay canvas no longer add opens actions to the Navigation stack

* Bug Fix

*Fixed issue where the game was not sending the Player to the correct
character conversation panel when pressing the "rollback" button,
causing the entire messaging app would close when it should not have.

* Messaging Bubble Update & Rollback Fix

*Fixed an issue that was preventing the rollback system from working
as expected (an extra node was being deleted by mistake) when the
selected choice from the DialogueChoiceNode was an action event.

*Updated the MessagingBubbles to increase the size of any media item
that are sent to the player.

*MessagingBubbles will now trim the surrounding box to fit the size of
the message content

+Added new variable on the DialogueUIManager that allows for easy
customisation of the maximum size/width of MessagingBubbles and media
content

* Messaging Panel & Scene Update

*Updated the Messaging Conversation Panel prefab - removing some unused
components.

*Updated main scene to fix some layout issues

* Notification Cleanup

*Cleaned up some issues with the notification prefab. Corrected some
anchor points and scaling issues that were present.

* Batched Notifications

+ Added batching logic and settings to DialogueUIManager. Notifications
from the same character are grouped so only one is shown at a time,
with an option to update the text to the latest message if desired. A
flag can also restrict the screen to a single notification, replacing
the previous one regardless of display time. Note: this system augments
batching and only applies when a new character sends a message.

* Navigation Bug Fix

*Fixed some issues related to the sorting order of canvas elements when
they were opened or closed using the navigation system.

*Fixed an issue where manually set thumbnails were not being used
inside the social media app, gallery canvas or messaging app